Coral Plot

Coral Plot

coral_plots(explainer, observation, grid_points = 101, select_points = 0.1)

Arguments

explainer
a model to be explained, preprocessed by the 'DALEX::explain' function

observation
a new observarvation for which predictions need to be explained

grid_points
number of points used for response path

select_points
number of points to present in coral plots

Value

An object of the class 'ceteris_paribus_explainer'. It's a data frame with calculated average responses.

Examples

library("DALEX")
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

new_apartment <- apartmentsTest[1, ]
new_apartment

#>      m2.price construction.year surface floor no.rooms    district
#> 1001     4644              1976     131     3        5 Srodmiescie

cr_rf <- coral_plots(explainer_rf, observation = new_apartment, select_points = 0.002)
cr_rf

#>         y_hat  new_x   vname quant obs_id        label
#> 1    4768.305  20.00 surface  0.00      0 randomForest
#> 2    4765.291  21.00 surface  0.01      0 randomForest
#> 3    4761.576  23.00 surface  0.02      0 randomForest
#> 4    4760.838  24.00 surface  0.03      0 randomForest
#> 5    4760.290  26.00 surface  0.04      0 randomForest
#> 6    4743.734  27.00 surface  0.05      0 randomForest
#> 7    4745.517  28.00 surface  0.06      0 randomForest
#> 8    4739.277  29.00 surface  0.07      0 randomForest
#> 9    4740.338  30.00 surface  0.08      0 randomForest
#> 10   4731.600  32.00 surface  0.09      0 randomForest
#> 11   4729.883  33.00 surface  0.10      0 randomForest
#> 12   4725.571  34.89 surface  0.11      0 randomForest
#> 13   4725.571  36.00 surface  0.12      0 randomForest
#> 14   4717.837  37.00 surface  0.13      0 randomForest
#> 15   4715.357  38.00 surface  0.14      0 randomForest
#> 16   4704.614  40.00 surface  0.15      0 randomForest
#> 17   4704.386  41.00 surface  0.16      0 randomForest
#> 18   4706.394  42.00 surface  0.17      0 randomForest
#> 19   4705.315  44.00 surface  0.18      0 randomForest
#> 20   4705.677  45.00 surface  0.19      0 randomForest
#> 21   4699.163  46.00 surface  0.20      0 randomForest
#> 22   4703.464  48.00 surface  0.21      0 randomForest
#> 23   4709.003  49.00 surface  0.22      0 randomForest
#> 24   4709.682  50.00 surface  0.23      0 randomForest
#> 25   4729.229  52.00 surface  0.24      0 randomForest
#> 26   4706.528  53.00 surface  0.25      0 randomForest
#> 27   4694.333  54.00 surface  0.26      0 randomForest
#> 28   4680.917  55.00 surface  0.27      0 randomForest
#> 29   4677.912  57.00 surface  0.28      0 randomForest
#> 30   4666.315  58.71 surface  0.29      0 randomForest
#> 31   4664.089  60.00 surface  0.30      0 randomForest
#> 32   4654.381  61.00 surface  0.31      0 randomForest
#> 33   4622.687  62.00 surface  0.32      0 randomForest
#> 34   4580.922  63.00 surface  0.33      0 randomForest
#> 35   4564.018  65.00 surface  0.34      0 randomForest
#> 36   4564.018  65.00 surface  0.35      0 randomForest
#> 37   4542.740  67.00 surface  0.36      0 randomForest
#> 38   4541.909  68.00 surface  0.37      0 randomForest
#> 39   4539.988  69.00 surface  0.38      0 randomForest
#> 40   4543.274  71.00 surface  0.39      0 randomForest
#> 41   4544.724  72.00 surface  0.40      0 randomForest
#> 42   4522.092  73.00 surface  0.41      0 randomForest
#> 43   4473.219  75.00 surface  0.42      0 randomForest
#> 44   4464.772  76.00 surface  0.43      0 randomForest
#> 45   4468.395  77.00 surface  0.44      0 randomForest
#> 46   4468.151  78.00 surface  0.45      0 randomForest
#> 47   4487.292  79.00 surface  0.46      0 randomForest
#> 48   4432.182  81.00 surface  0.47      0 randomForest
#> 49   4423.691  82.00 surface  0.48      0 randomForest
#> 50   4405.645  83.00 surface  0.49      0 randomForest
#> 51   4404.960  85.00 surface  0.50      0 randomForest
#> 52   4394.685  86.00 surface  0.51      0 randomForest
#> 53   4390.131  87.00 surface  0.52      0 randomForest
#> 54   4369.206  89.00 surface  0.53      0 randomForest
#> 55   4371.326  90.00 surface  0.54      0 randomForest
#> 56   4376.212  91.00 surface  0.55      0 randomForest
#> 57   4344.500  93.00 surface  0.56      0 randomForest
#> 58   4339.648  94.00 surface  0.57      0 randomForest
#> 59   4337.911  95.00 surface  0.58      0 randomForest
#> 60   4345.008  97.00 surface  0.59      0 randomForest
#> 61   4349.805  98.00 surface  0.60      0 randomForest
#> 62   4349.103  99.00 surface  0.61      0 randomForest
#> 63   4352.188 101.00 surface  0.62      0 randomForest
#> 64   4349.072 102.00 surface  0.63      0 randomForest
#> 65   4358.216 103.00 surface  0.64      0 randomForest
#> 66   4344.729 104.00 surface  0.65      0 randomForest
#> 67   4306.886 106.00 surface  0.66      0 randomForest
#> 68   4295.204 107.00 surface  0.67      0 randomForest
#> 69   4292.866 108.00 surface  0.68      0 randomForest
#> 70   4251.339 110.00 surface  0.69      0 randomForest
#> 71   4244.740 111.00 surface  0.70      0 randomForest
#> 72   4235.478 112.00 surface  0.71      0 randomForest
#> 73   4223.526 114.00 surface  0.72      0 randomForest
#> 74   4212.762 115.00 surface  0.73      0 randomForest
#> 75   4205.704 116.00 surface  0.74      0 randomForest
#> 76   4206.367 118.00 surface  0.75      0 randomForest
#> 77   4188.122 119.00 surface  0.76      0 randomForest
#> 78   4182.792 120.00 surface  0.77      0 randomForest
#> 79   4171.905 121.00 surface  0.78      0 randomForest
#> 80   4165.134 123.00 surface  0.79      0 randomForest
#> 81   4163.177 124.00 surface  0.80      0 randomForest
#> 82   4134.801 126.00 surface  0.81      0 randomForest
#> 83   4134.108 127.00 surface  0.82      0 randomForest
#> 84   4131.487 128.00 surface  0.83      0 randomForest
#> 85   4158.509 130.00 surface  0.84      0 randomForest
#> 86   4160.840 131.00 surface  0.85      0 randomForest
#> 87   4174.731 132.00 surface  0.86      0 randomForest
#> 88   4189.554 134.00 surface  0.87      0 randomForest
#> 89   4168.243 135.00 surface  0.88      0 randomForest
#> 90   4134.823 136.00 surface  0.89      0 randomForest
#> 91   4078.720 137.00 surface  0.90      0 randomForest
#> 92   4076.707 138.00 surface  0.91      0 randomForest
#> 93   4059.616 140.00 surface  0.92      0 randomForest
#> 94   4062.885 141.00 surface  0.93      0 randomForest
#> 95   4070.999 142.00 surface  0.94      0 randomForest
#> 96   4067.570 143.00 surface  0.95      0 randomForest
#> 97   4068.863 145.00 surface  0.96      0 randomForest
#> 98   4061.471 146.00 surface  0.97      0 randomForest
#> 99   4061.186 147.00 surface  0.98      0 randomForest
#> 100  4082.176 149.00 surface  0.99      0 randomForest
#> 101  4193.719 150.00 surface  1.00      0 randomForest
#> 102  4768.305  20.00 surface  0.00      1 randomForest
#> 103  4765.291  21.00 surface  0.01      1 randomForest
#> 104  4761.576  23.00 surface  0.02      1 randomForest
#> 105  4760.838  24.00 surface  0.03      1 randomForest
#> 106  4760.290  26.00 surface  0.04      1 randomForest
#> 107  4743.734  27.00 surface  0.05      1 randomForest
#> 108  4745.517  28.00 surface  0.06      1 randomForest
#> 109  4739.277  29.00 surface  0.07      1 randomForest
#> 110  4740.338  30.00 surface  0.08      1 randomForest
#> 111  4731.600  32.00 surface  0.09      1 randomForest
#> 112  4729.883  33.00 surface  0.10      1 randomForest
#> 113  4725.571  34.89 surface  0.11      1 randomForest
#> 114  4725.571  36.00 surface  0.12      1 randomForest
#> 115  4717.837  37.00 surface  0.13      1 randomForest
#> 116  4715.357  38.00 surface  0.14      1 randomForest
#> 117  4704.614  40.00 surface  0.15      1 randomForest
#> 118  4704.386  41.00 surface  0.16      1 randomForest
#> 119  4706.394  42.00 surface  0.17      1 randomForest
#> 120  4705.315  44.00 surface  0.18      1 randomForest
#> 121  4705.677  45.00 surface  0.19      1 randomForest
#> 122  4699.163  46.00 surface  0.20      1 randomForest
#> 123  4703.464  48.00 surface  0.21      1 randomForest
#> 124  4709.003  49.00 surface  0.22      1 randomForest
#> 125  4709.682  50.00 surface  0.23      1 randomForest
#> 126  4729.229  52.00 surface  0.24      1 randomForest
#> 127  4706.528  53.00 surface  0.25      1 randomForest
#> 128  4694.333  54.00 surface  0.26      1 randomForest
#> 129  4680.917  55.00 surface  0.27      1 randomForest
#> 130  4677.912  57.00 surface  0.28      1 randomForest
#> 131  4666.315  58.71 surface  0.29      1 randomForest
#> 132  4664.089  60.00 surface  0.30      1 randomForest
#> 133  4654.381  61.00 surface  0.31      1 randomForest
#> 134  4622.687  62.00 surface  0.32      1 randomForest
#> 135  4580.922  63.00 surface  0.33      1 randomForest
#> 136  4564.018  65.00 surface  0.34      1 randomForest
#> 137  4564.018  65.00 surface  0.35      1 randomForest
#> 138  4542.740  67.00 surface  0.36      1 randomForest
#> 139  4541.909  68.00 surface  0.37      1 randomForest
#> 140  4539.988  69.00 surface  0.38      1 randomForest
#> 141  4543.274  71.00 surface  0.39      1 randomForest
#> 142  4544.724  72.00 surface  0.40      1 randomForest
#> 143  4522.092  73.00 surface  0.41      1 randomForest
#> 144  4473.219  75.00 surface  0.42      1 randomForest
#> 145  4464.772  76.00 surface  0.43      1 randomForest
#> 146  4468.395  77.00 surface  0.44      1 randomForest
#> 147  4468.151  78.00 surface  0.45      1 randomForest
#> 148  4487.292  79.00 surface  0.46      1 randomForest
#> 149  4432.182  81.00 surface  0.47      1 randomForest
#> 150  4423.691  82.00 surface  0.48      1 randomForest
#> 151  4405.645  83.00 surface  0.49      1 randomForest
#> 152  4404.960  85.00 surface  0.50      1 randomForest
#> 153  4394.685  86.00 surface  0.51      1 randomForest
#> 154  4390.131  87.00 surface  0.52      1 randomForest
#> 155  4369.206  89.00 surface  0.53      1 randomForest
#> 156  4371.326  90.00 surface  0.54      1 randomForest
#> 157  4376.212  91.00 surface  0.55      1 randomForest
#> 158  4344.500  93.00 surface  0.56      1 randomForest
#> 159  4339.648  94.00 surface  0.57      1 randomForest
#> 160  4337.911  95.00 surface  0.58      1 randomForest
#> 161  4345.008  97.00 surface  0.59      1 randomForest
#> 162  4349.805  98.00 surface  0.60      1 randomForest
#> 163  4349.103  99.00 surface  0.61      1 randomForest
#> 164  4352.188 101.00 surface  0.62      1 randomForest
#> 165  4349.072 102.00 surface  0.63      1 randomForest
#> 166  4358.216 103.00 surface  0.64      1 randomForest
#>  [ reached getOption("max.print") -- omitted 1753 rows ]