Local Fit / Wangkardu Explanations

Local Fit / Wangkardu Explanations

local_fit(explainer, observation, selected_variable, grid_points = 101,
  select_points = 0.1)

Arguments

explainer
a model to be explained, preprocessed by the 'DALEX::explain' function

observation
a new observarvation for which predictions need to be explained

selected_variable
variable to be presented in the local fit plot

grid_points
number of points used for response path

select_points
fraction of points fromvalidation data to be presented in local fit plots

Value

An object of the class 'local_fit_explainer'. It's a data frame with calculated average responses.

Examples

library("DALEX")
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

new_apartment <- apartmentsTest[1, ]
new_apartment

#>      m2.price construction.year surface floor no.rooms    district
#> 1001     4644              1976     131     3        5 Srodmiescie

cr_rf <- local_fit(explainer_rf, observation = new_apartment,
    select_points = 0.002, selected_variable = "surface")
cr_rf

#>      y_hat new_x   vname quant obs_id        label
#> 1 4768.305    20 surface  0.00      0 randomForest
#> 2 4765.291    21 surface  0.01      0 randomForest
#> 3 4761.576    23 surface  0.02      0 randomForest
#> 4 4760.838    24 surface  0.03      0 randomForest
#> 5 4760.290    26 surface  0.04      0 randomForest
#> 6 4743.734    27 surface  0.05      0 randomForest